refactor: extract sensor descriptions and helpers into sensor_descriptions.py
